Character module improvement

+++ code/include/data.h
@@ -17,9 +17,9 @@ typedef struct Data{
     int direction;                                                   // Direction the player is facing
     char inventory[MAX_INVENTORY_SIZE][MAX_ITEM_NAME_LENGTH];        // Inventory of the player
     int inventory_count;                                             // Number of items in the inventory
-    bool picked_up_items[1];                                        // State of items in the world
+    bool picked_up_items[1];                                         // State of items in the world
 
-    int volume;                                                      // Volume of the game
+    float volume;            // Volume of the game
 } Data;
 
 Data LoadData(Settings* game_settings);                                          // Load game data
